# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Approval Workflow API",
        "description": "REST API Approval Workflow",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/approval-requests": {
            "get": {
                "tags": [
                    "Approval Requests"
                ],
                "summary": "Get Approval Requests",
                "description": "Menampilkan daftar approval request.",
                "operationId": "98c1d4ce5ffa8801d30da8513a7deed8",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter status",
                        "schema": {
                            "type": "string",
                            "example": "submitted"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari berdasarkan title",
                        "schema": {
                            "type": "string",
                            "example": "Laptop"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting data",
                        "schema": {
                            "type": "string",
                            "example": "latest"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Approval Requests"
                ],
                "summary": "Create Request",
                "operationId": "3015eac045c105f769fe57573e0ac026",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Pembelian Laptop"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Laptop untuk divisi IT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Request berhasil dibuat"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/approval-requests/{id}": {
            "get": {
                "tags": [
                    "Approval Requests"
                ],
                "summary": "Detail Request",
                "operationId": "e7ae51bd7540da946d14be7bdf704a9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Approval Requests"
                ],
                "summary": "Update Request",
                "operationId": "691a07361d86b6f0d234639d1d951ac1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Approval Requests"
                ],
                "summary": "Delete Request",
                "operationId": "b1d9b350424c1133421f28ae5beedc2a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Masukkan token dengan format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}